+++ src/ray.zig
@@ -12,7 +12,7 @@ pub const Ray = struct {
         };
     }
 
-    pub fn at(self: *const Ray, t: f32) vec3.Point3 {
+    pub fn at(self: *const Ray, t: f64) vec3.Point3 {
         return self.origin.add(self.direction.scalar_mul(t));
     }
 };
